==> 02_preprocessing.ipynb <==
import pandas as pd 
import numpy as np

#데이터 불러오기, 확인
df = pd.read_excel('../data/partyroom_A_review.xlsx')
df.head()

#id 비식별화
import hashlib

def anonymize_id(x):
    return hashlib.sha256(x.encode()).hexdigest()[:8]  # 8자리만 남기기

df['id'] = df['id'].apply(anonymize_id)
df.head()

#데이터 형태 확인
df.info()

#데이터 결측치 확인
df.isnull().sum()

#문자형 데이터 통계 확인
df.describe(include='O')

#결측치 처리
#visit_tm -> 새로운 카테고리(unknown)로 채우기
#content -> 새로운 카테고리(내용없음)로 채우기
print(df.isnull().sum())
df['visit_tm'] = df['visit_tm'].fillna('unknown')
df['content'] = df['content'].fillna('내용없음')
print(df.isnull().sum())

#df['keyword'] 형태 확인
print(type(df['keyword'][0]))

#df['keyword'] 내 공란('') 제거

import ast

cleaned_keywords = [] #행별 결과들을 모을 리스트

for row in df['keyword']:
    try:
        #문자열을 리스트로 변환
        keywords = ast.literal_eval(row) if isinstance(row, str) else row
    except(ValueError, SyntaxError):
        keywords = []
    
    #공백 제거 후 빈 문자열은 제외
    cleaned_row = [k.strip() for k in keywords if k.strip() != '']

    #비어있으면 '키워드 없음' 대체
    if not cleaned_row:
        cleaned_row = ['키워드 없음']

    cleaned_keywords.append(cleaned_row) #리스트 안의 리스트 구조 유지
        
df['keyword'] = cleaned_keywords
df['keyword']

#변경된 형태 확인
print(type(df['keyword'][0]))

# 전처리 완료 파일 저장
df2 = pd.DataFrame(df, columns=['id', 'content', 'visit_dt', 'visit_tm', 'visit_cnt', 'keyword'])
df2.to_excel('../data/A_preprocessing.xlsx', index=False)
