==> 03_analysis.ipynb <==
import pandas as pd
import numpy as np

#전처리 완료 파일 불러오기
df = pd.read_excel('../data/A_preprocessing.xlsx')
df.head()

-분석단계-<br>
1.파티룸 수요 분석<br>
2.충성도 지표(재방문) 분석<br>
3.키워드 분석<br>
4.이벤트유형 분석<br>
5.긍*부정 요인 분석

#파티룸 수요 분석

#visit_dt에서 월, 요일을 떼어내어 인기있는 월, 요일을 파악
df['month'] = df['visit_dt'].str.split().str[1]
df['date'] = df['visit_dt'].str.split().str[3]

#visit_dt 그룹화
mon_grouped = df.groupby(['month'])['visit_dt'].count().reset_index()
dt_grouped = df.groupby(['date'])['visit_dt'].count().reset_index()
tm_grouped = df.groupby(['visit_tm'])['visit_dt'].count().reset_index()

#그룹화하여 랭킹
hot_mon = mon_grouped.nlargest(5,'visit_dt') #월별 랭킹
hot_date = dt_grouped.nlargest(5,'visit_dt') #요일별 랭킹
tm_grouped #시간대별 

#비율(per) 추가
hot_mon['per'] = round(hot_mon['visit_dt']/len(df)*100,2)
hot_date['per'] = round(hot_date['visit_dt']/len(df)*100,2)
tm_grouped['per'] = round(tm_grouped['visit_dt']/len(df)*100,2)
tm_grouped


#충성도 지표(재방문) 분석

#리뷰 내용 중 재방문 관련 언급이 있는 리뷰 추출 / 휴리스틱 방식(직접 정의)
df['retention'] = df['content'].str.contains('또|재방문|늘|의사|다음|다시|단골|여기만')
df_true = df[df['retention']]

#재방문 언급률
#재방문 언급
true_cnt = len(df_true)

#전체
total_cnt = len(df)

#언급 비율
ratio = round(true_cnt / total_cnt *100, 2)

print(f'재방문 언급 리뷰 수 : {true_cnt}')
print(f'전체 리뷰 수 : {total_cnt}')
print(f'재방문 언급 비율 : {ratio}')

#리뷰 작성자별 방문 횟수
retention_grouped = df.groupby(['visit_cnt'])['id'].count().reset_index()
retention_grouped['per'] = round(retention_grouped['id']/len(df)*100, 2)
retention_grouped

#리뷰 작성자 중 실제 재방문률(A)
re_visit = retention_grouped.loc[1,'id'] + retention_grouped.loc[2,'id'] + retention_grouped.loc[3,'id']
#리뷰 작성자 중 실제 재방문률(B) 
# re_visit = retention_grouped.loc[1,'id']
retention_per = round(re_visit/len(df) * 100, 2)
print(f'실제 재방문률 : {retention_per}(리뷰수 기준)')

import ast
#문자열로 저장된 리스트를 실제 리스트로 변환
print(type(df['keyword'][0]))
df['keyword'] = df['keyword'].apply(lambda x: ast.literal_eval(x) if isinstance(x, str) else x)
print(type(df['keyword'][0]))

#키워드 분석
#df['keyword'] 행 안에 있는 리스트 들을 꺼내어 열 단위로 펼치기,'키워드 없음' 제외
df_exploded = df.explode('keyword') 
df_exploded = df_exploded[df_exploded['keyword'] != '키워드 없음']

#키워드 빈도 세기
keyword_fre = df_exploded['keyword'].value_counts().reset_index()
keyword_fre['per'] = round(keyword_fre['count']/keyword_fre['count'].sum()*100,2)
keyword_fre

#이벤트 유형 분석
#형태소 분석

from konlpy.tag import Okt
from collections import Counter
okt = Okt()

#df['content'] 행별 데이터 명사 추출
data = []
for row in df['content']:
    r_noun = okt.nouns(row)
    data.append(r_noun)

df['noun'] = data 
df['noun']

#이벤트 사전 만들기
event_map = {'생일':['생일', '생신', '생파', '생일파티', '환갑'],
             '회식':['회식', '회사', '동료', '친목'],
             '연말/연초모임':['연말모임','연초모임','연말','연초','신년','신년회','송년회'],
             '돌잔치':['돌잔치','백일','첫돌','돌', '잔치'],
             '크리스마스':['크리스마스','성탄','성탄절','xmas','트리', '산타'],
             '프로포즈':['프로포즈','고백','결혼'],
             '브라이덜샤워':['브샤','브라이덜샤워','브라이덜','샤워', '신부'],
             '동창회':['동창회','동창', '대학교', '고등학교'],
             '기타모임':['단체','동아리','소모임','동호회', '청첩장']}

#역매핑
event = {kw:evt for evt, kws in event_map.items() for kw in kws}
event_keys = set(event.keys())

#언급된 이벤트 별로 집계
evt_rv_cnt = Counter()

for tokens in df['noun']:
    mentioned = {event[t] for t in set(tokens) if t in event_keys} #event[t] 키워드를 대표 이벤트로 변환
    evt_rv_cnt.update(mentioned)

evt_rv_cnt

#전체 리뷰 중에서 이벤트성 방문 비율
evt_rank = pd.Series(evt_rv_cnt).sort_values(ascending=False)
evt_rank.sum() 
round(evt_rank.sum()/len(df)*100, 2)

#긍*부정 요인 분석(시설)
facility = ['노래방', '에어컨', '냉장고', '빔프로젝터', '화장실', '인테리어', '보드게임', '쇼파', '조명', '주차', 'PC방', '게임룸', '흡연실'] 

#긍정/부정 감성 단어 사전 
pos_word = ['장점', '갖춘', '추천', '강추', '좋', '좋았', '좋아', '좋더라', '좋습니', '좋고', '굿굿', '최고', '짱', '감탄', '최상', '만족', '깔끔', '깨끗', '깨끗하고', '빵빵', '시원', '쾌적', '이뻐', '이쁘고', '예뻐', '예뻤', '이쁨', '멋진', '힙해요', '매력', '재밌', '즐거웠', '즐겼' '넓고', '커서', '잘', '가성비','저렴', '괜찮', '편했', '편히', '편합', '아늑하고', '아늑한', '사진찍기', '포토존', '인생샷', '핫하', '딱이네', '빠르게', '신속하게', '매끄럽게', '빠르고', '빨라', '가능'] 
neg_word = ['단점', '빼고는', '아쉽', '아쉬웠', '아쉬워', '불만족', '별로', '글쎄', '더러워', '부족', '시끄럽', '시끄러', '비싸', '비싼', '짜증', '아깝', '아까웠', '좁다', '좁고', '협소', '협소한데', '습하고', '눅눅한', '눅눅', '불친절', '느리고', '느려', '답답', '힘들어', '힘들었', '어려웠' '주차불가', '불가', '외부에', '외부로', '바깥에', '밖에']


#.,!,? 기준으로 문장 분리
import re

def split_sentence(text):
    sentences = re.split(r'[.!?]|지만', text)
    return [s.strip() for s in sentences if s.strip()]

df['sentences'] = df['content'].apply(split_sentence)

#문장별 감성 분석 함수 생성

def classify_sentiment(sentence):
    if any(word in sentence for word in pos_word): 
        return '긍정'
    elif any(word in sentence for word in neg_word):
        return '부정'
    else:
        return '중립'

#시설별 감성 분석 결과 집계
facility_sentiment = Counter()

for sent_list in df['sentences']:
    for sent in sent_list:    
        sentiment = classify_sentiment(sent)
        for f in facility:
            if f in sent:
                facility_sentiment[(f, sentiment)] += 1

facility_sentiment

#피벗테이블
result = pd.DataFrame([(*k,v) for k,v in facility_sentiment.items()], columns = ['시설', '감정', '리뷰수'])
result_pivot = result.pivot(index='시설', columns='감정', values='리뷰수').fillna(0).astype(int)
result_pivot

#긍정 요인 랭킹
result_pivot['긍정'].sort_values(ascending=False)
